ParserService: Explicitly cast nbProcs to fit the smaller return value

+++ src/parser/ParserService.cpp
@@ -119,7 +119,7 @@ bool ParserService::isIdle() const
 
 uint8_t ParserService::nbNativeThreads() const
 {
-    auto nbProcs = compat::Thread::hardware_concurrency();
+    auto nbProcs = static_cast<uint8_t>( compat::Thread::hardware_concurrency() );
     if ( nbProcs == 0 )
         return 1;
     return nbProcs;
